Attempt use other http requests
