+++ src/main/java/ru/team/todo/controller/rest/UserRestController.java
@@ -4,34 +4,36 @@ import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-import ru.team.todo.dto.users.AddUserRequest;
-import ru.team.todo.dto.users.AddUserResponse;
-import ru.team.todo.dto.users.DeleteUserRequest;
-import ru.team.todo.dto.users.DeleteUserResponse;
+import org.springframework.web.bind.annotation.*;
+import ru.team.todo.dto.users.*;
 import ru.team.todo.services.UserService;
 
 import javax.validation.Valid;
 import java.util.HashMap;
 import java.util.Map;
 
-@RequestMapping("/action/user")
+@RequestMapping("/users")
 @RestController()
 @AllArgsConstructor
-public class UserActionRestController {
+public class UserRestController {
     private final UserService userService;
 
-    @PostMapping("/add")
+    @GetMapping
+    public FindUserResponse findAllUsers() {
+        return this.userService.findUser(new FindUserRequest(null));
+    }
+
+    @GetMapping("/{name}")
+    public FindUserResponse findUserByName(@PathVariable("name") String name) {
+        return this.userService.findUser(new FindUserRequest(name));
+    }
+
+    @PostMapping
     public AddUserResponse addUser(@RequestBody @Valid AddUserRequest request) {
         return this.userService.addUser(request);
     }
 
-    @PostMapping("/delete")
+    @DeleteMapping
     public DeleteUserResponse deleteUser(@RequestBody @Valid DeleteUserRequest request) {
         return this.userService.deleteUser(request);
     }
